backend/app/auth.py
import os
from datetime import datetime, timedelta, timezone
from typing import Annotated, Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from sqlalchemy.orm import Session
from passlib.context import CryptContext

from .database import get_db
from . import models, schemas

# --- Password Hashing Setup ---
# This configures passlib to use the bcrypt algorithm
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto") 

# --- JWT Configuration (from Environment Variables) ---
# Retrieve JWT settings from environment variables
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 120))

# Raise an error if SECRET_KEY is not set (critical for security)
if not SECRET_KEY:
    raise ValueError("SECRET_KEY environment variable is not set. Please set it in docker-compose.yml or your environment.")

# OAuth2PasswordBearer is used for handling token extraction from the request header
# It expects the token in the 'Authorization: Bearer <TOKEN>' header.
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token") # "token" is the endpoint for getting a token

# --- Password Utilities (moved from main.py for better organization, but still using pwd_context from main) ---
# Note: In a larger app, you might move pwd_context definition here or into a config file.
# For now, we're importing it from main.py as it's already there.

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verifies a plain password against a hashed password."""
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    """Hashes a plain password."""
    return pwd_context.hash(password)

# --- JWT Token Functions ---

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """
    Creates a new JWT access token.
    :param data: Data to encode into the token (e.g., user_id, username).
    :param expires_delta: Optional timedelta for expiration. If None, uses default.
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire}) # Add expiration time to the token payload
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def authenticate_user(db: Session, username: str, password: str):
    """
    Authenticates a user against the database.
    :returns: User object if authenticated, else None.
    """
    # Assuming 'username' can be either the actual username or email for login
    user = db.query(models.User).filter(
        (models.User.username == username) | (models.User.email == username)
    ).first()
    if not user:
        return None
    if not verify_password(password, user.hashed_password):
        return None
    return user

async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)], db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )

    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None:
            raise credentials_exception
        token_data = schemas.TokenData(username=username)

    except JWTError:
        raise credentials_exception

    user = db.query(models.User).filter(models.User.username == token_data.username).first()
    if user is None:
        raise credentials_exception
    return user