backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary<2.10
pydantic-settings
python-dotenv==1.0.1
sqlalchemy==2.0.30
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
httpx==0.27.0

# For OpenAI API integration
openai>=1.14.0

# For Sentence-Transformers embeddings
sentence-transformers==2.7.0

transformers==4.41.2
torch>=2.0.0
numpy==1.26.4